Final version

+++ main.cpp
@@ -8,7 +8,7 @@ int main(int argc,char *argv[])
 {
     string str = argv[1];
     int times = atoi(argv[2]);
-    double maxtime= atof(argv[3]);
+    int maxtime= atoi(argv[3]);
     int cnt=0;
     while(cnt<times)
     {
